default-values-and-collections: Restructure `impl` module
